# src/main/resources/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/servers/{serverId}": {
      "get": {
        "tags": [
          "ServerOperations"
        ],
        "summary": "Get server by id",
        "operationId": "getServer",
        "parameters": [
          {
            "name": "serverId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid id supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "Found the server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Server not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "ServerOperations"
        ],
        "summary": "Update a specific server",
        "operationId": "updateServer",
        "parameters": [
          {
            "name": "serverId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerUpdateRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid id supplied or Invalid fields supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "Server updated",
            "content": {
              "application/json": {

              }
            }
          },
          "404": {
            "description": "Server not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Server Error - Update Failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ServerOperations"
        ],
        "summary": "Delete a specific server",
        "operationId": "deleteServer",
        "parameters": [
          {
            "name": "serverId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid id supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "Deleted the server",
            "content": {
              "application/json": {

              }
            }
          },
          "404": {
            "description": "Server not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Server Error - Deletion Failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/servers": {
      "get": {
        "tags": [
          "ServerOperations"
        ],
        "summary": "Get all servers",
        "operationId": "getAllServers",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "Found servers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ServerOperations"
        ],
        "summary": "Add a new server",
        "operationId": "addServer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid fields supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "Server added",
            "content": {
              "application/json": {

              }
            }
          },
          "500": {
            "description": "Server Error - Insertion Failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/servers/upload": {
      "post": {
        "tags": [
          "ServerOperations"
        ],
        "operationId": "addServersFromFile",
        "parameters": [
          {
            "name": "file",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "binary"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/servers/downloadAll": {
      "get": {
        "tags": [
          "ServerOperations"
        ],
        "operationId": "downloadServers",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "path": {
            "type": "string"
          },
          "errorName": {
            "type": "string"
          },
          "reasons": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "errorMessage": {
            "type": "string"
          }
        }
      },
      "ServerUpdateRequestDto": {
        "type": "object",
        "properties": {
          "memory": {
            "type": "integer",
            "format": "int32",
            "minimum": 10
          },
          "noOfCPU": {
            "type": "integer",
            "format": "int32",
            "maximum": 9
          },
          "os": {
            "type": "string",
            "enum": [
              "WINDOWS",
              "LINUX",
              "UNIX"
            ]
          }
        }
      },
...
      "ServerRequestDto": {
        "type": "object",
        "properties": {
          "memory": {
            "type": "integer",
            "format": "int32",
            "minimum": 10
          },
          "noOfCPU": {
            "type": "integer",
            "format": "int32",
            "maximum": 9,
            "minimum": 1
          },
          "os": {
            "type": "string",
            "enum": [
              "WINDOWS",
              "LINUX",
              "UNIX"
            ]
          }
        },
        "required": [
          "memory",
          "noOfCPU",
          "os"
        ]
      },
...
      "ServerResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "datacenterId": {
            "type": "integer",
            "format": "int32"
          },
          "memory": {
            "type": "integer",
            "format": "int32"
          },
          "noOfCPU": {
            "type": "integer",
            "format": "int32"
          },
          "os": {
            "type": "string",
            "enum": [
              "WINDOWS",
              "LINUX",
              "UNIX"
            ]
          }
        }
      },
...
    }
  }
}